backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
langchain-core>=0.1.7,<0.2
openai>=1.6.1,<2.0.0
anthropic>=0.16.0,<1.0.0
pinecone-client==3.0.0
pinecone==3.0.0
tiktoken>=0.5.2,<0.6.0
httpx==0.25.2
python-dotenv==1.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-jaeger==1.21.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
pandas==2.1.3

